fnac/requirements.txt
# Dépendances pour le bot de connexion Fnac
selenium>=4.15.0
selenium-stealth>=1.0.6
undetected-chromedriver>=3.5.4
requests>=2.31.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.2

